сделал класс в отдельном файле
